Move data validation to separate module to stop complaints about pandas not found
