+++ tox.ini
@@ -42,7 +42,8 @@ commands =
 deps =
     -r{toxinidir}/f/common_logic/db_transformations.script.lock
     -r{toxinidir}/f/common_logic/db_operations.script.lock
-    -r{toxinidir}/f/common_logic/file_operations.script.lock
+    -r{toxinidir}/f/common_logic/save_disk.script.lock
+    -r{toxinidir}/f/common_logic/data_validation.script.lock
     -r{toxinidir}/f/common_logic/tests/requirements-test.txt
 commands =
     pytest {posargs} f/common_logic
